Iterators and constructors take two
Updated the iterators to be pointer types and switched to using
std::allocator for memory allocation.  (operator new doesn't work
because we have to explicitly construct and destruct elements.
Placement new would work alongside malloc, but std::allocator is there
so why not.)  Fixed up the constructors and now the constructor tests
pass.  (Still need to write the iterator constructor test.)